requirements.txt
Flask
pytesseract
Pillow
reportlab
Babel
opencv-python
numpy
